+++ application/src/main/java/org/opentripplanner/standalone/config/debuguiconfig/BackgroundTileLayer.java
@@ -1,0 +1,8 @@
+package org.opentripplanner.standalone.config.debuguiconfig;
+
+public record BackgroundTileLayer(
+  String name,
+  String templateUrl,
+  int tileSize,
+  String attribution
+) {}