+++ application/src/main/java/org/opentripplanner/street/model/edge/EscalatorEdge.java
@@ -1,5 +1,8 @@
 package org.opentripplanner.street.model.edge;
 
+import java.time.Duration;
+import java.util.Optional;
+import javax.annotation.Nullable;
 import org.locationtech.jts.geom.LineString;
 import org.opentripplanner.framework.geometry.GeometryUtils;
 import org.opentripplanner.framework.i18n.I18NString;
@@ -11,16 +14,14 @@ import org.opentripplanner.street.search.state.State;
 /** Represents an escalator. An escalator edge can only be traversed by walking */
 public class EscalatorEdge extends Edge {
 
-  /* A quick internet search gives escalator speed range of 0.3-0.6 m/s and angle of 30 degrees.
-   * Using the angle of 30 degrees and a speed of 0.5 m/s gives a horizontal component
-   * of approx. 0.43 m/s */
-  private static final double HORIZONTAL_SPEED = 0.45;
   private static final LocalizedString NAME = new LocalizedString("name.escalator");
   private final double length;
+  private final Duration duration;
 
-  private EscalatorEdge(Vertex v1, Vertex v2, double length) {
+  private EscalatorEdge(Vertex v1, Vertex v2, double length, Duration duration) {
     super(v1, v2);
     this.length = length;
+    this.duration = duration;
   }
 
   @Override
@@ -28,8 +29,13 @@ public class EscalatorEdge extends Edge {
     // Only allow traversal by walking
     if (s0.currentMode() == TraverseMode.WALK && !s0.getRequest().wheelchair()) {
       var s1 = s0.edit(this);
-      var time = getDistanceMeters() / HORIZONTAL_SPEED;
-      s1.incrementWeight(s0.getPreferences().walk().escalatorReluctance() * time);
+      double time;
+      if (duration == null) {
+        time = getDistanceMeters() / s0.getPreferences().walk().escalator().speed();
+      } else {
+        time = duration.toSeconds();
+      }
+      s1.incrementWeight(s0.getPreferences().walk().escalator().reluctance() * time);
       s1.incrementTimeInSeconds((int) Math.round(time));
       s1.incrementWalkDistance(getDistanceMeters());
       return s1.makeStateArray();
@@ -46,12 +52,25 @@ public class EscalatorEdge extends Edge {
     return length;
   }
 
+  /**
+   * Parsed content of duration tag in OSM, if any. Not a calculated value.
+   * @return Duration, or empty
+   */
+  public Optional<Duration> getDuration() {
+    return Optional.ofNullable(duration);
+  }
+
   @Override
   public I18NString getName() {
     return NAME;
   }
 
-  public static EscalatorEdge createEscalatorEdge(Vertex from, Vertex to, double length) {
-    return connectToGraph(new EscalatorEdge(from, to, length));
+  public static EscalatorEdge createEscalatorEdge(
+    Vertex from,
+    Vertex to,
+    double length,
+    @Nullable Duration duration
+  ) {
+    return connectToGraph(new EscalatorEdge(from, to, length, duration));
   }
 }