+++ application/src/main/java/org/opentripplanner/transit/model/basic/Money.java
@@ -148,6 +148,15 @@ public class Money implements Comparable<Money>, Serializable {
   }
 
   /**
+   * Returns the instance rounded down to the nearest multiple of 5 cents
+   * So $0.14 becomes $0.10
+   */
+  public Money roundDownToNearestFiveMinorUnits() {
+    int rounded = (this.minorUnitAmount() / 5) * 5;
+    return new Money(currency, rounded);
+  }
+
+  /**
    * Multiplies the amount with the multiplicator.
    */
   public Money times(int multiplicator) {